Added Invert, border-Style, Border Size, and Border Radius
Invert: Swaps so lower HP equals a fuller bar. To accommodate systems were health counts up to the max
Border Style: Adds a border around the bar using CSS. Left as a Text box to allow people to enter a style for each side of the box.
Border Size: How thick the border is
Border Radius: How curved the egdes of the border is.
Border Color:  What color each border is. Borders are for Forground and Background.